# Docs/OpenAPI.json
{
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Gateway service for Classroom microservices. Handles authentication, routing, and request aggregation.",
    "title": "Classroom Gateway API",
    "contact": {},
    "version": "1.0"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Создает новую учетную запись пользователя",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Регистрация пользователя",
        "parameters": [
          {
            "description": "Данные регистрации",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthRegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/AuthRegisterResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Пользователь уже существует",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Возвращает токены доступа и обновления",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Вход в систему",
        "parameters": [
          {
            "description": "Учетные данные",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AuthLoginResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Генерирует новую пару токенов по refresh-токену",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Обновление токена",
        "parameters": [
          {
            "description": "Refresh токен",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthRefreshRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AuthRefreshResponse"
            }
          },
          "401": {
            "description": "Неверный токен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Инвалидирует токены пользователя",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Выход из системы",
        "parameters": [
          {
            "description": "Данные для выхода",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthLogoutRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AuthLogoutResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/user-info": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает данные пользователя",
        "produces": ["application/json"],
        "tags": ["Auth"],
        "summary": "Информация о пользователе",
        "parameters": [
          {
            "type": "string",
            "example": "44e7f029-82cc-46f5-83e8-34b7d056ce32",
            "description": "Идентификатор пользователя",
            "name": "user_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AuthUserInfoResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает детальную информацию о курсе",
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Получение курса",
        "parameters": [
          {
            "type": "string",
            "example": "\"44e7f029-82cc-46f5-83e8-34b7d056ce32\"",
            "description": "ID курса",
            "name": "course_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCourseResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/create": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Создает новый курс",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Создание курса",
        "parameters": [
          {
            "description": "Данные для создания курса",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCourseRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateCourseResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course/enroll": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Записывает пользователя на указанный курс",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Запись на курс",
        "parameters": [
          {
            "description": "Данные для записи",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EnrollUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EnrollUserResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс или пользователь не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course/expel": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Отчисляет пользователя с указанного курса",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Отчисление с курса",
        "parameters": [
          {
            "description": "Данные для отчисления",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ExpelUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ExpelUserResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс или пользователь не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course/students": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список студентов, записанных на курс, с поддержкой пагинации",
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Получение студентов курса",
        "parameters": [
          {
            "type": "string",
            "example": "\"a3d8e9b0-5c1f-4e9d-8c1a-2b3c4d5e6f7a\"",
            "description": "ID курса",
            "name": "course_id",
            "in": "query",
            "required": true
          },
          {
            "minimum": 0,
            "type": "integer",
            "default": 0,
            "description": "Начальный индекс (для пагинации)",
            "name": "index",
            "in": "query",
            "required": true
          },
          {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "default": 10,
            "description": "Лимит записей (для пагинации)",
            "name": "limit",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCourseStudentsResponse"
            }
          },
          "400": {
            "description": "Некорректыные параметр",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course/update": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Обновляет информацию о курсе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Обновление курса",
        "parameters": [
          {
            "description": "Данные для обновления",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCourseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateCourseResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/course/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Удаляет курс по идентификатору",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Удаление курса",
        "parameters": [
          {
            "description": "Идентификатор курса",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteCourseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteCourseResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Курс не найден",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/courses": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список курсов с возможностью фильтрации по пользователю",
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Получение списка курсов",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCoursesResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/student-courses": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список курсов, на которые записан студент",
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Получение курсов студента",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCoursesResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/courses/teacher-courses": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Возвращает список курсов, которые ведет преподаватель",
        "produces": ["application/json"],
        "tags": ["Courses"],
        "summary": "Получение курсов преподавателя",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCoursesResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/lessons/create": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Создает новый урок в курсе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Lessons"],
        "summary": "Создание урока",
        "parameters": [
          {
            "description": "Данные для создания урока",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateLessonRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateLessonResponse"
            }
          },
          "400": {
            "description": "Некорректные данные",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Сервис недоступен",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AuthLogoutResponse": {
      "description": "Пустой ответ, указывающий на успешный выход",
      "type": "object"
    },
...
    "AuthRegisterRequest": {
      "description": "Содержит данные, необходимые для создания нового пользователя",
      "type": "object",
      "properties": {
        "email": {
          "description": "Email пользователя",
          "type": "string",
          "x-order": "0",
          "example": "user@example.com"
        },
        "first_name": {
          "description": "Имя пользователя",
          "type": "string",
          "x-order": "1",
          "example": "Иван"
        },
        "last_name": {
          "description": "Фамилия пользователя",
          "type": "string",
          "x-order": "2",
          "example": "Иванов"
        },
        "password": {
          "description": "Пароль",
          "type": "string",
          "x-order": "3",
          "example": "Qwerty123!"
        }
      }
    },
    "AuthRegisterResponse": {
      "description": "Возвращает ID созданного пользователя",
      "type": "object",
      "properties": {
        "user_id": {
          "description": "Уникальный идентификатор пользователя",
          "type": "string",
          "x-order": "0",
          "example": "d277084b-e1f6-4670-825b-53951d20b5d3"
        }
      }
    },
...
    "ErrorResponse": {
      "description": "Cтандартизированный формат всех ошибок в API.",
      "type": "object",
      "properties": {
        "code": {
          "description": "Код ошибки (HTTP статус код)",
          "type": "integer",
          "example": 400
        },
        "message": {
          "description": "Сообщение об ошибке",
          "type": "string",
          "example": "Bad request"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Enter JWT token in format: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}